% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_disagreement.R
\name{allocation_disagreement}
\alias{allocation_disagreement}
\title{Allocation Disagreement
Calculates the Allocation Disagreement between 2 matices}
\usage{
allocation_disagreement(p, q)
}
\arguments{
\item{p}{Reference Matrix}

\item{q}{Comparison Matrix}
}
\value{
Allocation Disagreement between the 2 matrices
}
\description{
Allocation Disagreement
Calculates the Allocation Disagreement between 2 matices
}

