% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_rel_area_overlap.R
\name{average_max_rel_area_overlap}
\alias{average_max_rel_area_overlap}
\title{Average Maximum Relative Area Overlap
Calculates the average maximum relative area overlap in a polygon set}
\usage{
average_max_rel_area_overlap(poly1, poly2)
}
\arguments{
\item{poly1}{type of SpatialPolygons, set of polygons for which the average maximum relative area overlap needs to be found.}

\item{poly2}{type of SpatialPolygons, set of polygons which needs to be overlapped with poly1.}
}
\value{
Average Maximum relative area overlap
}
\description{
Average Maximum Relative Area Overlap
Calculates the average maximum relative area overlap in a polygon set
}

