% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgecase_raster.r
\name{compR}
\alias{compR}
\title{Edge Cases
Checks all edge cases for rasters}
\usage{
compR(x, ..., extent = TRUE, rowcol = TRUE, crs = TRUE, res = TRUE,
  orig = TRUE, values = FALSE, stopiffalse = FALSE, showwarning = TRUE)
}
\arguments{
\item{x}{raster object}

\item{...}{raster objects}

\item{extent}{compares the bounding boxes}

\item{rowcol}{number of rows and columns are compared}

\item{crs}{coordinate refernece system are compared}

\item{res}{resolutions are compared}

\item{orig}{origins are compared}

\item{values}{cell values are compared}

\item{stopiffalse}{an error will occur}

\item{showwarning}{a warning is given}
}
\value{
True or False depending on if the edgecases pass or fail.
}
\description{
Edge Cases
Checks all edge cases for rasters
}

