% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{earthmd}
\alias{earthmd}
\title{Calculates Earth Movers Distance}
\usage{
earthmd(x, y, gc = FALSE, threshold = NULL)
}
\arguments{
\item{x}{A Spatial points object or a Raster}

\item{y}{a Spatial Points OBject or a Raster}

\item{gc}{True if great circle distances should be used.}

\item{threshold}{The maximal distance over which locations are compared.}
}
\value{
The earth movers distance between 2 spatial points data frames.
}
\description{
Takes in a 2 spatial points data frames and calculates the earth movers distance between them
}

