% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgecor_mapref.R
\name{edge.cor.ref}
\alias{edge.cor.ref}
\title{Edge Correction via map reflection}
\usage{
edge.cor.ref(ras, w)
}
\arguments{
\item{ras}{Given Raster}

\item{w}{weight}
}
\value{
iterative edge correction via map reflection
}
\description{
Edge Correction via map reflection
}

