% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc.R
\name{euc.dist}
\alias{euc.dist}
\title{Euclidean Distance
Calculates the Euclidean Distance between 2 spatial points}
\usage{
euc.dist(x1, x2)
}
\arguments{
\item{x}{spatial point}

\item{J}{spatial point}
}
\value{
Euclidean distance
}
\description{
Euclidean Distance
Calculates the Euclidean Distance between 2 spatial points
}

