% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_filter.r
\name{filter.g}
\alias{filter.g}
\title{Gaussian Filter Weights Matrix}
\usage{
filter.g(w, sigma)
}
\arguments{
\item{w}{weight}

\item{sigma}{cooeficient signma}
}
\value{
iterative edge correction via map reflection
}
\description{
Gaussian Filter Weights Matrix
}

