% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGauss.R
\name{getGauss}
\alias{getGauss}
\title{Get Gaussian Filter for raster object}
\usage{
getGauss(sigma, w)
}
\arguments{
\item{sigma}{is the standard deviation for the filter}

\item{w}{is the width of the neighbourhood in number of pixels out from centre cell}
}
\value{
\code{ssim} returns a filter object for use in \code{focal} function in the \code{raster} package
}
\description{
This function is a helper function for SSIM functions.
}
\details{
The \code{ssim} function can be used to generate the global multiscale ssim.
}
\seealso{
spatialcompare
}

