% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_rel_area_overlap.R
\name{maximum_rel_area_ovelap}
\alias{maximum_rel_area_ovelap}
\title{Calculates Maximum Relative Area Overlap}
\usage{
maximum_rel_area_ovelap(poly, p)
}
\arguments{
\item{poly}{is a set of polygons (as a Spatial Polygon Data Frame)}

\item{p}{is the polygon whose maximum relative area overlap needs to be determined.}
}
\value{
The maximum relative area overlap between the 2 polygon
}
\description{
Takes in a couple of polygons and calculates the relative area overlap between them
}

