% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msssim.R
\name{msssim}
\alias{msssim}
\title{Multiscale structural similarity index}
\usage{
msssim(img1, img2, w, gFIL = TRUE, edge = FALSE, ks = c(0.01, 0.03),
  level = 5, weight = c(0.0448, 0.2856, 0.3001, 0.2363, 0.1333),
  method = "product")
}
\arguments{
\item{img1}{is a \code{raster} object to compare}

\item{img2}{is a \code{raster} object to compare}

\item{w}{is the width of the neighbourhood in number of pixels out from centre cell}

\item{edge}{is a binary flag for whether a torroidal edge correction should be applied}

\item{ks}{is a vector of length 2 which contains values for constants in the SSIM formula. If ignored default values will be used.}

\item{GFil}{is a binary flag for whether a Gaussian filter should be applied as a smoothing function}
}
\value{
\code{msssim} returns a vector with the value of the multiscale structural similarity index
}
\description{
This function implements the multiscale structural similarity index of Wang et al. (2003) formulated
 for comparison of \code{raster} spatial objects at multiple scales.
}
\details{
The \code{msssim} function can be used to generate the global multiscale ssim. Local values are not supported. The weights used in the default are those given in Wang et al. (2003) and were determined empirically for a five-level application. The weights determine how cross-scale values of the SSIM are weighted.
}
\seealso{
ssim
}
\keyword{multi-scale}
\keyword{spatialcompare}

