% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_stats2.r
\name{O_G}
\alias{O_G}
\title{Overlay Statistics
OG - Mean distance from any location in the boundaries for G to the neareset location in the boundaries for H.}
\usage{
O_G(x, y)
}
\arguments{
\item{x}{raster1}

\item{J}{raster2}
}
\value{
O_G - Mean distance from any location in the boundaries for raster1 to the neareset location in the boundaries for raster2.
}
\description{
Overlay Statistics
OG - Mean distance from any location in the boundaries for G to the neareset location in the boundaries for H.
}

