% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_stats3.r
\name{O_GH}
\alias{O_GH}
\title{Overlay Statistics
OGH - Average distance from any location in the G or H boundaries to the neareset location in the boundaries for G.}
\usage{
O_GH(x, y)
}
\arguments{
\item{x}{raster1}

\item{J}{raster2}
}
\value{
O_GH - Average Distance from a location in raster1 or raster2 boundaries to the nearest location in the other(raster2 or raster1 as appropriate) boundary.
}
\description{
Overlay Statistics
OGH - Average distance from any location in the G or H boundaries to the neareset location in the boundaries for G.
}

