% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_stats1.r
\name{O_s}
\alias{O_s}
\title{Overlay Statistics
Os - Count of the number of locations that are in both the sets.}
\usage{
O_s(x, y)
}
\arguments{
\item{x}{raster1}

\item{J}{raster2}
}
\value{
O_s - Count of the number of locations that are in both the sets
}
\description{
Overlay Statistics
Os - Count of the number of locations that are in both the sets.
}

