% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_disagreement.R
\name{quantity_disagreement}
\alias{quantity_disagreement}
\title{Quantity Disagreement
Calculates the Quanitity Disagreement between 2 matices}
\usage{
quantity_disagreement(p, q)
}
\arguments{
\item{p}{Reference Matrix}

\item{q}{Comparison Matrix}
}
\value{
Quantity Disagreement between the 2 matrices
}
\description{
Quantity Disagreement
Calculates the Quanitity Disagreement between 2 matices
}

