% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_area_overlap.R
\name{relative_area_overlap}
\alias{relative_area_overlap}
\title{Relative Area Overlap}
\usage{
relative_area_overlap(p1, p2)
}
\arguments{
\item{p1}{The first spatial polygon (should be a list of coordinates)}

\item{p2}{The second spatial polygon (should be a list of coordinates)}
}
\value{
The relative area overlap between the 2 polygon
}
\description{
Takes in a couple of polygons and calucluates the relative area overlap between them
}

