% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simultaneous_area_overlap.R
\name{simultaneous_area_overlap}
\alias{simultaneous_area_overlap}
\title{Simultaneous area overlap
Calculates the Simultanoeus area overlap between 2 polygon sets}
\usage{
simultaneous_area_overlap(poly1, poly2)
}
\arguments{
\item{poly1}{type of SpatialPolygons}

\item{poly2}{type of SpatialPolygons}
}
\value{
Simultaneous area overlap
}
\description{
Simultaneous area overlap
Calculates the Simultanoeus area overlap between 2 polygon sets
}

