% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssim.R
\name{ssim}
\alias{ssim}
\title{Structural similarity index}
\usage{
ssim(img1, img2, w, gFIL = TRUE, edge = FALSE, ks = c(0.01, 0.03))
}
\arguments{
\item{img1}{is a \code{raster} object to compare}

\item{img2}{is a \code{raster} object to compare}

\item{w}{is the width of the neighbourhood in number of pixels out from centre cell}

\item{edge}{is a binary flag for whether a torroidal edge correction should be applied}

\item{ks}{is a vector of length 2 which contains values for constants in the SSIM formula. If ignored default values will be used.}

\item{GFil}{is a binary flag for whether a Gaussian filter should be applied as a smoothing function}
}
\value{
\code{ssim} returns a list with 5 objects. The first is the value of the global ssim statistic. Objects 2-5 are all \code{raster} objects
  that contain the composite index, luminance, contrast, and structure components respectively.
}
\description{
This function implements the structural similarity index of Wang et al. (2003) formulated
 for comparison of \code{raster} spatial objects.
}
\details{
The \code{ssim} function can be used to generate the three component indices of the SSIM index
 as well as the composite index. Two raster objects are the key arguments which will be used to compare.
 These two raster objects must have the same \code{CRS}, extent, and cell size.
}
\seealso{
spatialcompare
}
\keyword{hello}
\keyword{spatialcompare}

