% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_decomposition.r
\name{wavelet_decomposition}
\alias{wavelet_decomposition}
\title{Wavelet Decomposition
Applies the Two-Dimensional Discrete Haar Wavelet Filter}
\usage{
wavelet_decomposition(x, J = 4)
}
\arguments{
\item{x}{input matrix (image)}

\item{J}{depth of decomposition}
}
\value{
List structure containing the 3J + 1 sub-matrices from the decomposition
}
\description{
Wavelet Decomposition
Applies the Two-Dimensional Discrete Haar Wavelet Filter
}

