% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_deomposition2.r
\name{wavelet_decomposition2}
\alias{wavelet_decomposition2}
\title{Wavelet Decomposition
Applies the Two-Dimensional Discrete Haar Wavelet Filter}
\usage{
wavelet_decomposition2(y)
}
\arguments{
\item{y}{object returned by dwt}
}
\value{
List structure containing the 3J + 1 sub-matrices from the decomposition
}
\description{
Wavelet Decomposition
Applies the Two-Dimensional Discrete Haar Wavelet Filter
}

