% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_average_rel_overlap.R
\name{weighted_avg_maximum_relative_area_overlap}
\alias{weighted_avg_maximum_relative_area_overlap}
\title{Weighted Average Area Overlap
Calculates the Weighted Average Maximum Relative Area Overlap of a polygon set}
\usage{
weighted_avg_maximum_relative_area_overlap(poly1, poly2)
}
\arguments{
\item{poly1}{type of SpatialPolygons, set of polygons for which the weighted average maximum relative area overlap needs to be found.}

\item{poly2}{type of SpatialPolygons, set of polygons which needs to be overlapped with poly1.}
}
\value{
Weighted Average Maximum Relative Area Overlap
}
\description{
Weighted Average Area Overlap
Calculates the Weighted Average Maximum Relative Area Overlap of a polygon set
}

