% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_simultaneous.R
\name{weighted_simultanous_area_overlap}
\alias{weighted_simultanous_area_overlap}
\title{Weighted Simultaneous area overlap
Calculates the Weighted Simultanoeus area overlap between 2 polygon sets}
\usage{
weighted_simultanous_area_overlap(poly1, poly2)
}
\arguments{
\item{poly1}{type of SpatialPolygons}

\item{poly2}{type of SpatialPolygons}
}
\value{
Weighted Simultaneous area overlap
}
\description{
Weighted Simultaneous area overlap
Calculates the Weighted Simultanoeus area overlap between 2 polygon sets
}

